Move NotFoundText to typography
