+++ packages/ui/src/common/components/typography/NotFoundText.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components'
 
-import { TextBig } from './typography'
+import { TextBig } from '.'
 
 export const NotFoundText = styled(TextBig).attrs({ lighter: true })`
   justify-self: center;
