+++ packages/ui/src/forum/components/threads/ThreadList.tsx
@@ -4,7 +4,7 @@ import styled from 'styled-components'
 import { List } from '@/common/components/List'
 import { ListHeader, ListHeaders } from '@/common/components/List/ListHeader'
 import { Loading } from '@/common/components/Loading'
-import { NotFoundText } from '@/common/components/NotFoundText'
+import { NotFoundText } from '@/common/components/typography/NotFoundText'
 import { RowGapBlock } from '@/common/components/page/PageContent'
 import { HeaderText, SortIconDown, SortIconUp } from '@/common/components/SortedListHeaders'
 import { ThreadsColLayout } from '@/forum/constant'
