+++ packages/ui/src/proposals/components/ProposalList/ProposalList.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 
 import { List } from '@/common/components/List'
 import { ListHeader } from '@/common/components/List/ListHeader'
-import { NotFoundText } from '@/common/components/NotFoundText'
+import { NotFoundText } from '@/common/components/typography/NotFoundText'
 import { RowGapBlock } from '@/common/components/page/PageContent'
 import { ProposalColLayout, ProposalsListHeaders } from '@/proposals/constants'
 import { Proposal } from '@/proposals/types'
