+++ packages/ui/src/working-groups/components/OpeningsList/LoadingOpenings.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 
 import { Loading } from '@/common/components/Loading'
-import { NotFoundText } from '@/common/components/NotFoundText'
+import { NotFoundText } from '@/common/components/typography/NotFoundText'
 import { OpeningsList } from '@/working-groups/components/OpeningsList/OpeningsList'
 import { UpcomingWorkingGroupOpening, WorkingGroupOpening } from '@/working-groups/types'
